Keep CELT prefilter scratch float-width

+++ celt/accessors.go
@@ -85,15 +85,15 @@ func (d *Decoder) NextRNG() uint32 {
 }
 
 // GetEnergy returns the energy for a specific band and channel from prevEnergy.
-func (d *Decoder) GetEnergy(band, channel int) float64 {
+func (d *Decoder) GetEnergy(band, channel int) float32 {
 	if band < 0 || band >= MaxBands || channel < 0 || channel >= d.channels {
 		return 0
 	}
-	return float64(d.prevEnergy[channel*MaxBands+band])
+	return float32(d.prevEnergy[channel*MaxBands+band])
 }
 
 // SetEnergy sets the energy for a specific band and channel.
-func (d *Decoder) SetEnergy(band, channel int, energy float64) {
+func (d *Decoder) SetEnergy(band, channel int, energy float32) {
 	if band < 0 || band >= MaxBands || channel < 0 || channel >= d.channels {
 		return
 	}
